somnus_bot/somnus_tg_bot/requirements.txt
aiofiles           22.1.0
aiogram            3.0.0b6
aiohttp            3.8.4
aioschedule        0.5.2
aiosignal          1.3.1
async-timeout      4.0.2
attrs              23.1.0
certifi            2022.12.7
charset-normalizer 3.1.0
frozenlist         1.3.3
idna               3.4
magic-filter       1.0.9
multidict          6.0.4
pip                22.0.2
psycopg2-binary    2.9.6
pydantic           1.10.9
python-dotenv      1.0.0
redis              4.5.5
requests           2.31.0
setuptools         59.6.0
typing_extensions  4.6.3
urllib3            2.0.3
yarl               1.9.2